src/requirements/production.txt
# taggit through model doesn't seem to work with django4, try again later
#Django==4.0.4
Django==4.1.1
django-allauth==0.51.0
django-htmx==1.12.2
django-taggit==3.0.0
django-polymorphic==3.1.0
django-ninja==0.19.1
environs[django]==9.5.0
orjson==3.8.0
psycopg2-binary==2.9.3
Pillow==9.1.1
django-imagekit==4.1.0
django-bootstrap-v5==1.0.11
fontawesomefree==6.2.0
dealer==2.1.0
python-magic==0.4.27
